Move foundational components into namespace
