+++ src/components/Things/Scroller.svelte
@@ -1,7 +1,5 @@
 <script>
-    import Card from "../Card.svelte";
-    import Image from "../Image.svelte";
-    import Text from "../Text.svelte";
+    import { Card, Image, Text } from "../foundation";
 
     export let things = [];
     things = things.sort((a, b) => b.stock - a.stock);
