+++ src/components/foundation/index.js
@@ -1,0 +1,8 @@
+export { default as Card } from './Card.svelte';
+export { default as Column } from './Column.svelte';
+export { default as Image } from './Image.svelte';
+export { default as Link } from './Link.svelte';
+export { default as Row } from './Row.svelte';
+export { default as Text } from './Text.svelte';
+export { default as TextInput } from './TextInput.svelte';
+export { default as Title } from './Title.svelte';
